+++ src/main/kotlin/it/pagopa/helpdeskcommands/repositories/TransactionsEventStoreRepository.kt
@@ -1,4 +1,4 @@
-package it.pagopa.helpdeskcommands.dataproviders.repositories.ecommerce
+package it.pagopa.helpdeskcommands.repositories
 
 import it.pagopa.ecommerce.commons.documents.BaseTransactionEvent
 import org.springframework.data.mongodb.repository.ReactiveMongoRepository
